Added DefaultGame parameter to Condition.isFulfilled method, and moved RequirementCondition class to conditions package
