+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/condition/Condition.java
@@ -1,5 +1,7 @@
 package com.gempukku.stccg.condition;
 
+import com.gempukku.stccg.game.DefaultGame;
+
 public interface Condition {
-    boolean isFulfilled();
+    boolean isFulfilled(DefaultGame cardGame);
 }
