+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/condition/HigherScoreThanAllOtherPlayersCondition.java
@@ -12,9 +12,8 @@ public class HigherScoreThanAllOtherPlayersCondition implements Condition {
     }
 
     @Override
-    public boolean isFulfilled() {
+    public boolean isFulfilled(DefaultGame cardGame) {
         Player targetPlayer = _playerResolver.getPlayer();
-        DefaultGame cardGame = targetPlayer.getGame();
         for (Player player : cardGame.getPlayers()) {
             if (player.getScore() >= targetPlayer.getScore() && targetPlayer != player) {
                 return false;
