+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/condition/TrueCondition.java
@@ -1,6 +1,8 @@
 package com.gempukku.stccg.condition;
 
+import com.gempukku.stccg.game.DefaultGame;
+
 public class TrueCondition implements Condition {
     @Override
-    public boolean isFulfilled() { return true; }
+    public boolean isFulfilled(DefaultGame cardGame) { return true; }
 }
